requirements.txt
# Core
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0

# Media processing
pymediainfo==6.1.0
pillow==10.1.0

# Database
sqlalchemy==2.0.23

# Web framework
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0

# HTTP clients
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
humanize==4.9.0

# NZB handling
pynzb==0.1.0

# File hashing
xxhash==3.4.1
